requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
psycopg2-binary
WTForms
WTForms[sqlalchemy]
email-validator
torch
wfdb
numpy
werkzeug
